+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 - Add ``lower_first``. Thanks bharadwajyarlagadda_!
 - Add ``upper_case``. Thanks bharadwajyarlagadda_!
 - Add ``upper_first``. Thanks bharadwajyarlagadda_!
+- Add ``subtract``. Thanks bharadwajyarlagadda_!
 
 
 v3.4.4 (2016-09-06)